scripts/requirements_professional.txt
torch>=1.9.0
transformers>=4.20.0
gradio>=3.0.0
sentencepiece>=0.1.96
protobuf>=3.20.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
datasets>=2.0.0
evaluate>=0.4.0
sacrebleu>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.0.0
